Moved some integration-test-only code to integration tests folder
